use submodule for metrics
