+++ bb8-metrics/Dockerfile
@@ -1,7 +1,9 @@
 FROM tiangolo/uwsgi-nginx-flask:python3.6
 
 COPY requirements.txt .
+COPY app/metrics/requirements.txt ./app/metrics/
 RUN pip3 install -r requirements.txt
+RUN pip3 install -r app/metrics/requirements.txt
 
 # Install docker
 RUN apt-get update
