+++ bb8-metrics/app/static/index.html
